requirements.txt
fastapi
uvicorn
python-multipart
groq
PyMuPDF
sentence-transformers
numpy
python-dotenv
httpx
